moved the item placeholder creation to its own class removed unused imports
